requirements.txt
requests==2.32.3
icalendar==6.1.3
recurring-ical-events==3.6.1
google-api-python-client==2.167.0
google-auth-httplib2==0.2.0 
google-auth-oauthlib==1.2.1
google-cloud-storage==2.12.0
pandas==2.1.3